# docs/openapi.yml
openapi: 3.0.1
info:
  title: My Library API
  description: REST API for My Library
  version: v1
paths:
  /api/books:
    post:
      tags:
        - Books
      summary: Adds a new book to the library.
      operationId: AddBookAsync
      requestBody:
        description: The request containing new book details.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBookRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/AddBookRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/AddBookRequest"
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    get:
      tags:
        - Books
      summary: Gets all books in the library.
      operationId: GetAllBooksAsync
      responses:
        "200":
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponse"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponse"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponse"
        "401":
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  "/api/books/{id}/borrow":
    post:
      tags:
        - Books
      summary: Borrows a book for the logged in user.
      operationId: BorrowBookAsync
      parameters:
        - name: id
          in: path
          description: The book ID
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        description: The request.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowBookRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/BorrowBookRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/BorrowBookRequest"
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Server Error
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /api/profile:
    post:
      tags:
        - Customers
      summary: Updates the logged in user's profile.
      operationId: UpdateProfileAsync
      requestBody:
        description: The request.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
components:
  schemas:
    AddBookRequest:
      required:
        - id
        - title
      type: object
      properties:
        id:
          type: string
          description: "Required: The book ID."
          format: uuid
          example: EABAA778-F8F2-4491-B80D-F9B5B15A0237
        title:
          minLength: 1
          type: string
          description: "Required: The book title."
          example: The Green Moon by John Doe
        numberOfCopies:
          maximum: 1000
          minimum: 1
          type: integer
          description: The number of copies. Should be between 1 and 1000.
          format: int32
          example: 18
      additionalProperties: false
    BookResponse:
      required:
        - id
        - title
      type: object
      properties:
        id:
          type: string
          description: The book id.
          format: uuid
          example: EABAA778-F8F2-4491-B80D-F9B5B15A0237
        title:
          minLength: 1
          type: string
          description: The book title.
          example: The Green Moon by John Doe
        availableCopies:
          type: integer
          description: Total copies available in the library at the moment.
          format: int32
          example: 17
        nextAvailableCopyDate:
          type: string
          description: The closest date when a borrowed copy will be returned.
          format: date
          nullable: true
          example: "2023-12-25"
      additionalProperties: false
    BorrowBookRequest:
      type: object
      properties:
        returnDate:
          type: string
          description: The date on which the customer expects to return the book.
          format: date
          example: "2023-12-25"
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: {}
...
    UpdateProfileRequest:
      required:
        - emailAddress
        - fullName
      type: object
      properties:
        fullName:
          minLength: 1
          type: string
          description: The first and last name.
          example: John Doe
        emailAddress:
          minLength: 1
          type: string
          description: The email address.
          format: email
          example: john@email.com
      additionalProperties: false
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://vinarah.eu.auth0.com/authorize?audience=api://my-library.co.za
          scopes:
            openid: Open Id
security:
  - oauth2:
      - openid